# dev: optimize string handling and slice allocations