Improve performance in the goformatters internal `diff` implementation by:
- Using byte operations instead of string conversions
- Preallocating slices with exact capacity
- Replacing `slices.Concat` with manual appends

Benchmarks show faster line parsing, less memory usage and fewer memory allocations per operation.

Current `main`:

```
BenchmarkParseDiffLines-8        2649880               430.3 ns/op           880 B/op         11 allocs/op
BenchmarkHunkChangesParser-8     1414388               864.1 ns/op          1424 B/op         23 allocs/op
BenchmarkLargeDiff-8                4202            292897 ns/op          822887 B/op       7039 allocs/op
```

Proposed changes:

```
BenchmarkParseDiffLines-8        3590509               327.1 ns/op           560 B/op          8 allocs/op
BenchmarkHunkChangesParser-8     1596944               741.9 ns/op          1104 B/op         20 allocs/op
BenchmarkLargeDiff-8                4635            263006 ns/op          495908 B/op       7015 allocs/op
```