@ldez 2025-02-04
The gains are very low.

Can you share your bench test code? (not inside the code of the PR but as a comment)