@ldez 2025-02-04
Micro-benchmarks stay micro-benchmarks, in the real usage of golangci-lint, there is no advantage.

```console
$ hyperfine --warmup 1 \
-n 'master' --prepare "./golangci-lint-prev cache clean" "./golangci-lint-prev run --issues-exit-code 0 --print-issued-lines=false --enable-only goimports" \
-n 'local' --prepare './golangci-lint-new cache clean' "./golangci-lint-new run --issues-exit-code 0 --print-issued-lines=false --enable-only goimports"
Benchmark 1: master
  Time (mean ± σ):      1.060 s ±  0.030 s    [User: 2.416 s, System: 1.478 s]
  Range (min … max):    1.010 s …  1.105 s    10 runs
 
Benchmark 2: local
  Time (mean ± σ):      1.094 s ±  0.060 s    [User: 2.485 s, System: 1.473 s]
  Range (min … max):    1.035 s …  1.200 s    10 runs
 
Summary
  master ran
    1.03 ± 0.06 times faster than local
```


```console
$ ./golangci-lint-prev run --issues-exit-code 0 --print-issued-lines=false --enable-only goimports -v
INFO golangci-lint has version (devel) built with go1.23.5 from (89b146e04bbfb180036ccb270ab4d05928115194, modified: false, mod sum: "") on 2025-02-04T14:06:02Z 
INFO [config_reader] Config search paths: [./ /home/ldez/sources/golangci/golangci-lint /home/ldez/sources/golangci /home/ldez/sources /home/ldez /home /] 
INFO [config_reader] Used config file .golangci.yml 
INFO [goenv] Read go env for 2.113ms: map[string]string{"GOCACHE":"/home/ldez/.cache/go-build", "GOROOT":"/usr/lib/go"} 
INFO [lintersdb] Active 1 linters: [goimports]    
INFO [loader] Go packages loading at mode 8199 (name|compiled_files|files) took 651.083202ms 
INFO [runner/filename_unadjuster] Pre-built 0 adjustments in 10.946744ms 
INFO [linters_context/goanalysis] analyzers took 2.763894375s with top 10 stages: goimports: 2.762746916s, typecheck: 1.147459ms 
INFO [runner/skip_dirs] Skipped 2 issues from dir internal/go/cache by pattern internal/go 
INFO [runner] Issues before processing: 2, after processing: 0 
INFO [runner] Processors filtering stat (in/out): path_relativity: 2/2, path_absoluter: 2/2, invalid_issue: 2/2, exclusion_paths: 2/2, skip_dirs: 2/0, skip_files: 2/2, cgo: 2/2, filename_unadjuster: 2/2 
INFO [runner] processing took 28.776µs with stages: skip_dirs: 19.082µs, path_relativity: 2.679µs, invalid_issue: 1.066µs, cgo: 757ns, skip_files: 706ns, nolint_filter: 605ns, path_absoluter: 499ns, max_same_issues: 457ns, sort_results: 372ns, filename_unadjuster: 365ns, fixer: 327ns, uniq_by_line: 308ns, max_from_linter: 217ns, diff: 204ns, exclusion_paths: 181ns, exclusion_rules: 179ns, path_shortener: 178ns, generated_file_filter: 127ns, identifier_marker: 110ns, source_code: 107ns, max_per_file_from_linter: 85ns, path_prettifier: 83ns, severity-rules: 82ns 
INFO [runner] linters took 254.666367ms with stages: goanalysis_metalinter: 254.568835ms 
INFO File cache stats: 0 entries of total size 0B 
INFO Memory: 11 samples, avg is 49.8MB, max is 81.3MB 
INFO Execution took 921.039093ms   
```

```console
$ ./golangci-lint-new run --issues-exit-code 0 --print-issued-lines=false --enable-only goimports -v
INFO golangci-lint has version (devel) built with go1.23.5 from (fd63603ef936cff4a24b8a78c731e1dbaa7d36ba, modified: true, mod sum: "") on 2025-02-04T19:56:09Z 
INFO [config_reader] Config search paths: [./ /home/ldez/sources/golangci/golangci-lint /home/ldez/sources/golangci /home/ldez/sources /home/ldez /home /] 
INFO [config_reader] Used config file .golangci.yml 
INFO [goenv] Read go env for 2.213953ms: map[string]string{"GOCACHE":"/home/ldez/.cache/go-build", "GOROOT":"/usr/lib/go"} 
INFO [lintersdb] Active 1 linters: [goimports]    
INFO [loader] Go packages loading at mode 8199 (compiled_files|files|name) took 759.496894ms 
INFO [runner/filename_unadjuster] Pre-built 0 adjustments in 29.113546ms 
INFO [linters_context/goanalysis] analyzers took 2.802974897s with top 10 stages: goimports: 2.80107387s, typecheck: 1.901027ms 
INFO [runner/skip_dirs] Skipped 2 issues from dir internal/go/cache by pattern internal/go 
INFO [runner] Issues before processing: 2, after processing: 0 
INFO [runner] Processors filtering stat (in/out): path_absoluter: 2/2, path_relativity: 2/2, exclusion_paths: 2/2, cgo: 2/2, filename_unadjuster: 2/2, skip_files: 2/2, invalid_issue: 2/2, skip_dirs: 2/0 
INFO [runner] processing took 31.73µs with stages: skip_dirs: 18.766µs, path_relativity: 4.297µs, invalid_issue: 1.16µs, skip_files: 840ns, cgo: 838ns, max_same_issues: 702ns, nolint_filter: 674ns, path_absoluter: 609ns, filename_unadjuster: 577ns, sort_results: 426ns, fixer: 402ns, generated_file_filter: 293ns, exclusion_paths: 268ns, diff: 253ns, exclusion_rules: 245ns, path_shortener: 206ns, source_code: 204ns, uniq_by_line: 198ns, severity-rules: 185ns, max_from_linter: 177ns, identifier_marker: 155ns, max_per_file_from_linter: 133ns, path_prettifier: 122ns 
INFO [runner] linters took 287.078373ms with stages: goanalysis_metalinter: 286.988691ms 
INFO File cache stats: 0 entries of total size 0B 
INFO Memory: 12 samples, avg is 49.4MB, max is 81.8MB 
INFO Execution took 1.079660972s  
```