@ldez 2025-02-04
The "complexity" introduced to replace `slices.Concat` and the truncation isn't worth the cost.
The gains are very low and it decreases maintainability.

So I will decline the PR, thank you anyway.